pyproject.toml
[project]
name = "slm"
version = "0.1.0"
description = "SLM iterative optimization algorithms for holographic beam shaping and tweezer array generation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
app = ["streamlit>=1.30"]
jax = ["jax>=0.4", "jaxlib>=0.4"]
hardware = [
    "wxPython>=4.2",
    "vmbpy",
    "Pillow>=9.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py312"
